+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.2
+
+* InstallmentCount Added for transaction functions.
+* CardHolderName added for 3D init function.
+
 ## 1.0.1
 
 * Reverse function added.